+++ changelog.md
@@ -93,6 +93,10 @@
 0.14
 - RPCS3 is working from SteamUI thanks to Flatseal
 
+0.15
+- Complete revamp of the interface
+- Now you can customize some RetroArch Options: Bezels
+- Emulators can now be skipped so you can mantain your custom configuration
 
-0.14.1
+0.15.1
 - Added support for .rpx Wii U Games